# specs/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Stellar Scout API",
		"version": "1.6.1",
		"summary": "Agent-native data layer for the Stellar ecosystem — projects, builders, hackathons, SCF, audits, research.",
...
		"contact": {
			"name": "Stellar Light",
			"url": "https://stellarlight.xyz/scout",
			"email": "support@stellarlight.xyz"
		},
		"license": {
			"name": "MIT",
			"identifier": "MIT"
		}
	},
	"servers": [
		{
			"url": "https://stellarlight.xyz",
			"description": "Production"
		}
	],
...
	"paths": {
...
		"/api/changelog": {
			"get": {
				"operationId": "getChangelog",
				"tags": [
					"Discovery"
				],
				"summary": "Recent changes to the API, MCP tools, and client",
				"description": "A curated, latest-first feed of contract-affecting changes — new/removed endpoints & tools, param/enum changes, description rewrites. Point an agent here to see what changed lately without reading git history. **Use when:** you cached the API surface earlier and want to know what moved before relying on it, or you're debugging a field/param that changed. **Not for:** the actual ecosystem data → use the relevant search endpoint; current health / source freshness / version → use /api/status.",
				"parameters": [
					{
						"name": "since",
						"in": "query",
						"required": false,
						"description": "Only entries on/after this ISO date (YYYY-MM-DD).",
						"schema": {
							"type": "string",
							"format": "date"
						}
					},
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"description": "Max entries to return, latest-first (1–100).",
						"schema": {
							"type": "integer",
							"minimum": 1,
							"maximum": 100
						}
					}
				],
				"responses": {
					"200": {
						"description": "Changelog feed (latest-first)",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"ok": {
											"type": "boolean"
										},
										"service": {
											"type": "string"
										},
										"version": {
											"type": "string"
										},
										"generatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"meta": {
											"type": "object"
										},
										"entries": {
											"type": "array",
											"items": {
												"type": "object",
												"properties": {
													"date": {
														"type": "string",
														"format": "date"
													},
													"surfaces": {
														"type": "array",
														"items": {
															"type": "string",
															"enum": [
																"api",
																"mcp",
																"api-client",
																"skill"
															]
														}
													},
													"version": {
														"type": "string"
													},
													"type": {
														"type": "string",
														"enum": [
															"added",
															"changed",
															"fixed",
															"removed"
														]
													},
													"summary": {
														"type": "string"
													},
													"detail": {
														"type": "string"
													}
												}
											}
										}
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/api/research": {
			"get": {
				"operationId": "searchResearch",
				"tags": [
					"Research"
				],
				"summary": "Vector search over the Stellar research corpus",
				"description": "Semantic ($vectorSearch over Voyage embeddings) search across the Stellar **knowledge corpus** — SDF blog, SCF Handbook, SEPs/standards, dev docs, papers, SCF proposals, security audits, incident reports, and Lumenloop/EC research. Returns the top matching text chunks with source attribution, section, URL, and a confidence score (relevance + freshness + authority); audit chunks add auditor/protocol/severity. Filterable by `source`; falls back to BM25-lite keyword search if vectors are unavailable. **Use when:** 'how does X work', 'is X possible / has X been discussed on Stellar', 'what does the SEP/spec/audit say about X', or you need primary-source citations for a thesis or design question. **Not for:** what products exist or their funding/status → use search_projects; GitHub source code ranked by quality → use search_repos. Use for ecosystem KNOWLEDGE & explainer questions — 'what is / how does / who / why / is it true that…'. Topics: security & risk (bug bounty, vulnerability disclosure, security audit, exploit, hack, incident, post-mortem, oracle manipulation, YieldBlox, Reflector); compliance & regulation (Travel Rule, FATF, KYC, AML, BSA, sanctions, humanitarian aid); funding & governance (SCF v7.0, award tiers, grant programs, Neural Quorum Governance, community voting, quorum); the Stellar Development Foundation (SDF) — mission, legal structure, org; protocol history (Stellar Consensus Protocol, SCP, whitepaper, authors); ecosystem programs (ambassador program, regional chapters, India, bootcamps); and fact-checking or verifying a claim about Stellar. More topics this corpus answers: SCF program mechanics — how to apply (step-by-step application process), Build Award, Instawards, Liquidity Award (amounts), Public Goods Award, eligibility, review timeline (prescreen, panel review, community vote, KYC), submission deadlines; Soroban security incidents — reentrancy, soroban-sdk security advisories/CVEs, denial-of-service (instance-storage growth, authorization recursion); ecosystem history — Soroban mainnet launch (Protocol 20), XLM initial supply creation and distribution, humanitarian aid disbursements (UNHCR Ukraine), SDF Enterprise Fund investments, SDF mandate/spending priorities; making an issued asset tradable/visible on exchanges, wallets, explorers, and aggregators; contract source verification (release.yml, Stellar Lab vs stellar.expert).",
				"parameters": [
					{
						"name": "q",
						"in": "query",
						"required": true,
						"description": "Natural-language search query",
						"schema": {
							"type": "string",
							"minLength": 2
						}
					},
					{
						"name": "source",
						"in": "query",
						"description": "Optional source filter. Use 'audit' for security questions, 'ec-developer-report' for ecosystem stats, 'paper' for foundational protocol questions.",
						"schema": {
							"type": "string",
							"enum": [
								"sdf-blog",
								"scf-handbook",
								"sep",
								"dev-docs",
								"paper",
								"scf-proposal",
								"lumenloop",
								"lumenloop-research",
								"audit",
								"incident",
								"ec-developer-report"
							]
						}
					},
					{
						"name": "limit",
						"in": "query",
						"description": "Max results (default 8, max 25)",
						"schema": {
							"type": "integer",
							"minimum": 1,
							"maximum": 25,
							"default": 8
						}
					}
				],
				"responses": {
					"200": {
						"description": "Research results",
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								}
							}
						}
					}
				}
			}
		},
...
		"/api/feedback": {
			"get": {
				"operationId": "getFeedbackSchema",
				"tags": [
					"Feedback"
				],
				"summary": "Get the feedback request schema (discovery)",
				"description": "Returns the expected POST body shape + valid `kind` values, so an agent can self-discover how to submit feedback without guessing. No side effects. **Use when:** you're about to POST feedback and want the schema first — kind enum, required fields, size limits. **Not for:** actually submitting feedback → POST /api/feedback; browsing what's been reported → not exposed (curator queue is private).",
				"responses": {
					"200": {
						"description": "Feedback request schema",
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								}
							}
						}
					}
				}
			},
			"post": {
				"operationId": "submitFeedback",
				"x-side-effecting": true,
				"tags": [
					"Feedback"
				],
				"summary": "Submit feedback on Scout's output",
				"description": "Send a feedback report when the skill returns wrong / missing / misleading information. Lands in the curator queue. Rate-limited to 6/min/IP (IP hashed with PAYLOAD_SECRET, never stored raw). **Use when:** a Scout query returned something wrong / stale / empty that you believe SHOULD exist — file it so the corpus/endpoint gets fixed. **Not for:** discovering the schema first → GET /api/feedback; reading data (this is a write endpoint) → use the relevant search endpoint.",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/FeedbackRequest"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Feedback received (a curator-queue row was created)",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"required": [
										"ok",
										"id"
									],
									"properties": {
										"ok": {
											"type": "boolean"
										},
										"id": {
											"type": "string"
										},
										"message": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Validation error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"429": {
						"description": "Rate limit exceeded",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		}
	},
	"components": {
...
		"schemas": {
...
			"ErrorResponse": {
				"type": "object",
				"required": [
					"error"
				],
				"properties": {
					"error": {
						"type": "string"
					},
					"details": {
						"type": "object",
						"additionalProperties": true
					}
				}
			},
...
			"FeedbackRequest": {
				"type": "object",
				"required": [
					"kind",
					"message"
				],
				"description": "Optional reporting context is NESTED under `context` (matches the live endpoint + the GET /api/feedback self-schema).",
				"properties": {
					"kind": {
						"type": "string",
						"enum": [
							"bug",
							"missing-data",
							"wrong-answer",
							"suggestion",
							"other"
						]
					},
					"message": {
						"type": "string",
						"minLength": 10,
						"maxLength": 4000
					},
					"context": {
						"type": "object",
						"description": "Optional context about what triggered the feedback.",
						"properties": {
							"query": {
								"type": "string"
							},
							"endpoint": {
								"type": "string"
							},
							"skillVersion": {
								"type": "string"
							},
							"agentName": {
								"type": "string"
							}
						}
					}
				}
			}
		}
	}
}